requirements.txt
numpy

matplotlib 

scipy

ipykernel

PyWavelets

python-dotenv